rag/requirements.txt
flask
flask-cors
requests
faiss-cpu
psycopg2-binary
pytest

